requirements.txt
# FastAPI and dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
python-multipart>=0.0.6
starlette>=0.27.0

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.6

# Authentication and security
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1

# AWS
boto3>=1.28.0

# Document processing
pdf2image>=1.16.3
pytesseract>=0.3.10
Pillow>=10.0.0

# OpenAI
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Testing
pytest>=7.3.1
httpx>=0.24.1